improve time format
